[Frontend] [Middleware] [Backend] run Simulation + Graphs (excluding the stack bar) (#83)
* Updated: visualize Chart uses SimulatioId and not scenarioId

* UNSTABLE Modified: ensure simulation can run
if state tax of married is null, it won't run (for single user)
updated the execution path of the child threads

* Updated: delete a null scenario will just exit

* UNSTABLE Modified: temporary solution for what we have
user have a previousSimulation object over the 2 other variable which we will use in later hw

* UNSTABLE Updated: run simulation prototype
not final, but at least it works in terms of middleware

* Updated: corrected variable naming

* Modified: removed uneeded variable

* Modified: ensure an id is attached to each graph added

* UNSTABLE Updated: view line chart
the algorithm assume no simulation failed

* Updated: shaded line chart render order / color opacity

* Updated: Shaded line chart

* Updated: only show chart if no error

* Updated: make setChart disappear on wrong

* Updated: all state YAML file to be decimal rate

* Updated: added a WA state tax
the fallback tax as no state tax

* Updated: selection of tax logic

* Modified: removed the null check on state tax
as every state tax default to WA if none is created as the 0 state tax

* Updated: added the scenario name to the page

* Updated: chart permission

* Updated: better way to obtain the path of runWorker

* Patched: error pass in catch method

+++ server/db/controllers/SimulationController.js
@@ -33,6 +33,16 @@ export default class SimulationController {
         }
     }
 
+    async read(id) {
+        try {
+            const simulation = await Simulation.findById(id).populate("results");
+            return simulation;
+        }
+        catch (error) {
+            throw new Error(error);
+        }
+    }
+
     /**
      * This function delete the Simulation with the given id
      * @note
@@ -48,6 +58,9 @@ export default class SimulationController {
         const resultController = new ResultController();
         try {
             const simulation = await Simulation.findById(id);
+            if (!simulation) {
+                return;
+            }
             for (const result of simulation.results) {
                 await resultController.delete(result);
             }
@@ -57,7 +70,7 @@ export default class SimulationController {
             throw new Error(error);
         }
     }
-    
+
     /**
      * This function deletes a Simulation with the given id
      * @param {mongoose.Types.ObjectId} id
